+++ src/js/repeat.js
@@ -32,16 +32,20 @@ export default {
 
         $repeatInfos = this.form.view.$.find( '.or-repeat-info' );
         this.templates = {};
-        // Add repeat numbering elements, if repeat has form controls (not just calculations)
-        $repeatInfos.siblings( '.or-repeat' )
+        // Add repeat numbering elements
+        $repeatInfos
+            .siblings( '.or-repeat' )
+            .prepend( '<span class="repeat-number"></span>' )
+            // add empty class for calculation-only repeats
+            .addBack()
             .filter( function() {
-                // remove whitespace so we can use :empty css selector
+                // remove whitespace
                 if ( this.firstChild && this.firstChild.nodeType === 3 ) {
                     this.firstChild.textContent = '';
                 }
-                return !!this.querySelector( '.question' );
+                return !this.querySelector( '.question' );
             } )
-            .prepend( '<span class="repeat-number"></span>' );
+            .addClass( 'empty' );
         // Add repeat buttons
         $repeatInfos.filter( '*:not([data-repeat-fixed]):not([data-repeat-count])' )
             .append( '<button type="button" class="btn btn-default add-repeat-btn"><i class="icon icon-plus"> </i></button>' )