+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+--------------------
+##### Fixed
+- Failing to clone repeats that only contain calculations.
+
 [5.2.2] - 2019-04-01
 --------------------
 ##### Changed